Create a new file for all opcode inspection functions.
