+++ source/opt/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_library(SpvToolsOpt
   ${CMAKE_CURRENT_SOURCE_DIR}/constructs.h
   ${CMAKE_CURRENT_SOURCE_DIR}/def_use_manager.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/reflect.h
   ${CMAKE_CURRENT_SOURCE_DIR}/spv_builder.h
   ${CMAKE_CURRENT_SOURCE_DIR}/types.h
   ${CMAKE_CURRENT_SOURCE_DIR}/type_builder.h
